# docs/swagger.yaml
basePath: /api
definitions:
  api.JsonError:
    properties:
      message:
        example: error message
        type: string
    type: object
...
  service.Genre:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: horror
        type: string
    type: object
  service.GenreData:
    properties:
      name:
        maxLength: 255
        type: string
    required:
    - name
    type: object
  service.Movie:
    properties:
      duration_minutes:
        example: 255
        type: integer
      id:
        example: 1
        type: integer
      poster_url:
        example: https://www.amazon.nl/Lord-Strings-J-Khan/dp/B000BU9TAQ
        type: string
      released_at:
        example: "2025-10-30T00:00:00Z"
        type: string
      synopsis:
        example: A man that can manipulate every string variable
        type: string
      title:
        example: Lord of the Strings
        type: string
    type: object
  service.MovieData:
    properties:
      duration_minutes:
        type: integer
      poster_url:
        type: string
      released_at:
        type: string
      synopsis:
        maxLength: 5000
        type: string
      title:
        maxLength: 255
        type: string
    required:
    - duration_minutes
    - released_at
    - synopsis
    - title
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Esta es la API REST para la aplicación de gestión de películas.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: API de Películas UNICEN
  version: "1.0"
paths:
...
  /genres:
    get:
      consumes:
      - application/json
      description: Obtiene una lista de géneros, ordenada por fecha de estreno descendente,
        con paginación.
      parameters:
      - description: Número de página
        in: query
        name: page
        required: true
        type: integer
      - description: Cantidad de filas por página
        in: query
        name: rows
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de géneros exitosa
          schema:
            items:
              $ref: '#/definitions/service.Genre'
            type: array
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Lista géneros paginados
      tags:
      - Genres
    post:
      consumes:
      - application/json
      description: Agrega un nuevo género al catálogo.
      parameters:
      - description: Datos del género
        in: body
        name: genre
        required: true
        schema:
          $ref: '#/definitions/service.GenreData'
      produces:
      - application/json
      responses:
        "201":
          description: Género agregado exitosamente
          schema:
            $ref: '#/definitions/service.Genre'
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "409":
          description: Error de duplicacion de recursos
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Agrega un género
      tags:
      - Genres
  /genres/{ID}:
    delete:
      consumes:
      - application/json
      description: Borra un género dado por ID permanentemente del catálogo.
      parameters:
      - description: ID del género a eliminar
        in: path
        name: ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Género eliminado exitosamente
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "404":
          description: Error por género no encontrado
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Borra un género
      tags:
      - Genres
    get:
      consumes:
      - application/json
      description: Obtiene toda la información asociada a un género dado por ID.
      parameters:
      - description: ID del género a obtener
        in: path
        name: ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Género obtenida exitosamente
          schema:
            $ref: '#/definitions/service.Genre'
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "404":
          description: Error por género no encontrado
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Obtiene un género
      tags:
      - Genres
    put:
      consumes:
      - application/json
      description: Actualiza toda la información de un género dado por ID. Todos los
        campos son requeridos.
      parameters:
      - description: ID del género a actualizar
        in: path
        name: ID
        required: true
        type: integer
      - description: Datos actualizables del género
        in: body
        name: genre
        required: true
        schema:
          $ref: '#/definitions/service.GenreData'
      produces:
      - application/json
      responses:
        "200":
          description: Género actualizado exitosamente
          schema:
            $ref: '#/definitions/service.Genre'
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "404":
          description: Error por género no encontrado
          schema:
            $ref: '#/definitions/api.JsonError'
        "409":
          description: Error de duplicación de recursos
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Actualiza los datos de un género
      tags:
      - Genres
...
  /movies:
    get:
      consumes:
      - application/json
      description: Obtiene una lista de películas, ordenada por fecha de estreno descendente,
        con paginación.
      parameters:
      - description: Número de página
        in: query
        name: page
        required: true
        type: integer
      - description: Cantidad de filas por página
        in: query
        name: rows
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de películas exitosa
          schema:
            items:
              $ref: '#/definitions/service.Movie'
            type: array
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Lista películas paginadas
      tags:
      - Movies
    post:
      consumes:
      - application/json
      description: Agrega una nueva película al catálogo. La URL del póster puede
        omitirse.
      parameters:
      - description: Datos de la película
        in: body
        name: movie
        required: true
        schema:
          $ref: '#/definitions/service.MovieData'
      produces:
      - application/json
      responses:
        "201":
          description: Película agregada exitosamente
          schema:
            $ref: '#/definitions/service.Movie'
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "409":
          description: Error de duplicacion de recursos
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Agrega una película
      tags:
      - Movies
  /movies/{ID}:
    delete:
      consumes:
      - application/json
      description: Borra una película dada por ID permanentemente del catálogo.
      parameters:
      - description: ID de la película a eliminar
        in: path
        name: ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Película eliminada exitosamente
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "404":
          description: Error por película no encontrada
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Borra una película
      tags:
      - Movies
    get:
      consumes:
      - application/json
      description: Obtiene toda la información asociada a una película dada por ID.
      parameters:
      - description: ID de la película a obtener
        in: path
        name: ID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Película obtenida exitosamente
          schema:
            $ref: '#/definitions/service.Movie'
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "404":
          description: Error por película no encontrada
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Obtiene una película
      tags:
      - Movies
    put:
      consumes:
      - application/json
      description: Actualiza toda la información de una película dada por ID. Todos
        los campos son requeridos.
      parameters:
      - description: ID de la película a actualizar
        in: path
        name: ID
        required: true
        type: integer
      - description: Datos actualizables de la película
        in: body
        name: movie
        required: true
        schema:
          $ref: '#/definitions/service.MovieData'
      produces:
      - application/json
      responses:
        "200":
          description: Película actualizada exitosamente
          schema:
            $ref: '#/definitions/service.Movie'
        "400":
          description: Error de validación de los parámetros
          schema:
            $ref: '#/definitions/api.JsonError'
        "404":
          description: Error por película no encontrada
          schema:
            $ref: '#/definitions/api.JsonError'
        "409":
          description: Error de duplicación de recursos
          schema:
            $ref: '#/definitions/api.JsonError'
        "500":
          description: Error interno del servidor
          schema:
            $ref: '#/definitions/api.JsonError'
      summary: Actualiza los datos de una película
      tags:
      - Movies
...
swagger: "2.0"